Move Edwards and Montgomary to their own modules
