+++ ed448-goldilocks/src/curve.rs
@@ -1,9 +1,2 @@
-pub mod edwards;
-pub mod montgomery;
-mod scalar;
 pub(crate) mod scalar_mul;
 pub(crate) mod twedwards;
-
-pub use edwards::{AffinePoint, CompressedEdwardsY, EdwardsPoint};
-pub use montgomery::{MontgomeryPoint, ProjectiveMontgomeryPoint};
-pub use scalar::{EdwardsScalar, EdwardsScalarBytes, WideEdwardsScalarBytes};
