+++ ed448-goldilocks/src/edwards.rs
@@ -11,5 +11,7 @@
 /// If this is a problem, one can use a different isogeny strategy (Decaf/Ristretto)
 pub(crate) mod affine;
 pub(crate) mod extended;
+mod scalar;
 pub use affine::AffinePoint;
 pub use extended::{CompressedEdwardsY, EdwardsPoint};
+pub use scalar::{EdwardsScalar, EdwardsScalarBytes, WideEdwardsScalarBytes};
